+++ packages/design-system/src/lib/components/form/form-group/form-element/FormFeedback.tsx
@@ -1,21 +1,21 @@
 import { FormControl } from 'react-bootstrap'
 import { PropsWithChildren } from 'react'
-import { Col } from '../../../grid/Col'
+import { Col, ColProps } from '../../../grid/Col'
+import { Row } from '../../../grid/Row'
 
-interface FormFeedbackProps {
+interface FormFeedbackProps extends ColProps {
   type?: 'valid' | 'invalid'
-  withinMultipleFieldsGroup?: boolean
+  as?: typeof Col | typeof Row
 }
 
 export function FormFeedback({
   type = 'valid',
-  withinMultipleFieldsGroup,
-  children
+  as,
+  children,
+  ...props
 }: PropsWithChildren<FormFeedbackProps>) {
-  return withinMultipleFieldsGroup ? (
-    <FormControl.Feedback type={type}>{children}</FormControl.Feedback>
-  ) : (
-    <FormControl.Feedback as={Col} sm={{ offset: 3 }} type={type}>
+  return (
+    <FormControl.Feedback type={type} as={as} {...props}>
       {children}
     </FormControl.Feedback>
   )