+++ src/sections/create-dataset/CreateDatasetFactory.tsx
@@ -1,24 +1,29 @@
 import { ReactElement } from 'react'
 import { useSearchParams } from 'react-router-dom'
-import { CreateDatasetForm } from './CreateDatasetForm'
+import { CreateDataset } from './CreateDataset'
 import { DatasetJSDataverseRepository } from '../../dataset/infrastructure/repositories/DatasetJSDataverseRepository'
 import { MetadataBlockInfoJSDataverseRepository } from '../../metadata-block-info/infrastructure/repositories/MetadataBlockInfoJSDataverseRepository'
+import { NotImplementedModalProvider } from '../not-implemented/NotImplementedModalProvider'
 
 const repository = new DatasetJSDataverseRepository()
 const metadataBlockInfoRepository = new MetadataBlockInfoJSDataverseRepository()
 
 export class CreateDatasetFactory {
   static create(): ReactElement {
-    return <CreateDatasetFormWithSearchParams />
+    return (
+      <NotImplementedModalProvider>
+        <CreateDatasetWithSearchParams />
+      </NotImplementedModalProvider>
+    )
   }
 }
 
-function CreateDatasetFormWithSearchParams() {
+function CreateDatasetWithSearchParams() {
   const [searchParams] = useSearchParams()
   const collectionId = searchParams.get('collectionId') ?? undefined
 
   return (
-    <CreateDatasetForm
+    <CreateDataset
       repository={repository}
       metadataBlockInfoRepository={metadataBlockInfoRepository}
       collectionId={collectionId}