+++ src/sections/create-dataset/MetadataBlockFormFields/MetadataFormField/Fields/Vocabulary.tsx
@@ -1,24 +1,71 @@
-import { Form } from '@iqss/dataverse-design-system'
-import { forwardRef } from 'react'
+import { useMemo } from 'react'
+import { Controller, useFormContext } from 'react-hook-form'
+import { Form, Row, Col } from '@iqss/dataverse-design-system'
+import { type CommonFieldProps } from '..'
+import { MetadataFieldsHelper } from '../../../MetadataFieldsHelper'
 
-interface Props {
+interface VocabularyProps extends CommonFieldProps {
+  metadataBlockName: string
   options: string[]
-  onChange: (event: React.ChangeEvent<HTMLSelectElement>) => void
-  isInvalid: boolean
+  compoundParentName?: string
+  fieldsArrayIndex?: number
 }
-// TODO: Implement a multiple select with autocomplete like in JSF version
-export const Vocabulary = forwardRef(function Vocabulary(
-  { options, onChange, isInvalid, ...props }: Props,
-  ref
-) {
-  return (
-    <Form.Group.Select onChange={onChange} isInvalid={isInvalid} ref={ref} {...props}>
-      <option value="">Select</option>
-      {options.map((option) => (
-        <option key={option} value={option}>
-          {option}
-        </option>
-      ))}
-    </Form.Group.Select>
+export const Vocabulary = ({
+  name,
+  compoundParentName,
+  metadataBlockName,
+  rulesToApply,
+  description,
+  title,
+  options,
+  isRequired,
+  withinMultipleFieldsGroup,
+  fieldsArrayIndex
+}: VocabularyProps) => {
+  const { control } = useFormContext()
+
+  const builtFieldName = useMemo(
+    () =>
+      MetadataFieldsHelper.defineFieldName(
+        name,
+        metadataBlockName,
+        compoundParentName,
+        fieldsArrayIndex
+      ),
+    [name, metadataBlockName, compoundParentName, fieldsArrayIndex]
   )
-})
+
+  return (
+    <Controller
+      name={builtFieldName}
+      control={control}
+      rules={rulesToApply}
+      render={({ field: { onChange, ref }, fieldState: { invalid, error } }) => (
+        <Form.Group controlId={builtFieldName} as={withinMultipleFieldsGroup ? Col : Row}>
+          <Form.Group.Label
+            message={description}
+            required={isRequired}
+            column={!withinMultipleFieldsGroup}
+            sm={3}>
+            {title}
+          </Form.Group.Label>
+          <Col sm={withinMultipleFieldsGroup ? 12 : 9}>
+            <Row>
+              <Col sm={withinMultipleFieldsGroup ? 12 : 9}>
+                <Form.Group.Select onChange={onChange} isInvalid={invalid} ref={ref}>
+                  <option value="">Select</option>
+                  {options.map((option) => (
+                    <option key={option} value={option}>
+                      {option}
+                    </option>
+                  ))}
+                </Form.Group.Select>
+                <Form.Group.Feedback type="invalid">{error?.message}</Form.Group.Feedback>
+              </Col>
+            </Row>
+          </Col>
+        </Form.Group>
+      )}
+    />
+  )
+}