+++ src/sections/create-dataset/useCreateDatasetForm.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { createDataset } from '../../dataset/domain/useCases/createDataset'
 import { DatasetRepository } from '../../dataset/domain/repositories/DatasetRepository'
-import { MetadataFieldsHelper } from './MetadataFieldsHelper'
+import { type CreateDatasetFormValues, MetadataFieldsHelper } from './MetadataFieldsHelper'
 import { Route } from '../Route.enum'
 
 export enum SubmissionStatus {
@@ -12,32 +12,28 @@ export enum SubmissionStatus {
   Errored = 'Errored'
 }
 
-export type FormCollectedValues = Record<
-  string,
-  Record<string, string | string[] | FormCollectedComposedFields>
->
-export type FormCollectedComposedFields = Record<string, string>
-
-export function useCreateDatasetForm(repository: DatasetRepository): {
+export function useCreateDatasetForm(
+  repository: DatasetRepository,
+  collectionId: string
+): {
   submissionStatus: SubmissionStatus
-  submitForm: (formData: FormCollectedValues) => void
+  submitForm: (formData: CreateDatasetFormValues) => void
 } {
   const [submissionStatus, setSubmissionStatus] = useState<SubmissionStatus>(
     SubmissionStatus.NotSubmitted
   )
   const navigate = useNavigate()
 
-  const submitForm = (formData: FormCollectedValues): void => {
+  const submitForm = (formData: CreateDatasetFormValues): void => {
     setSubmissionStatus(SubmissionStatus.IsSubmitting)
 
-    const formDataBackToOriginalKeys = MetadataFieldsHelper.replaceSlashKeysWithDot(
-      formData
-    ) as FormCollectedValues
+    const formDataBackToOriginalKeys = MetadataFieldsHelper.replaceSlashKeysWithDot(formData)
+
     const formattedFormValues = MetadataFieldsHelper.formatFormValuesToCreateDatasetDTO(
       formDataBackToOriginalKeys
     )
 
-    createDataset(repository, formattedFormValues)
+    createDataset(repository, formattedFormValues, collectionId)
       .then(({ persistentId }) => {
         setSubmissionStatus(SubmissionStatus.SubmitComplete)
         navigate(`${Route.DATASETS}?persistentId=${persistentId}`, {