+++ src/stories/create-dataset/CreateDataset.stories.tsx
@@ -1,32 +1,39 @@
 import type { StoryObj, Meta } from '@storybook/react'
-import { CreateDatasetForm } from '../../sections/create-dataset/CreateDatasetForm'
+import { CreateDataset } from '../../sections/create-dataset/CreateDataset'
 import { WithLayout } from '../WithLayout'
 import { WithI18next } from '../WithI18next'
 import { DatasetMockRepository } from '../dataset/DatasetMockRepository'
 import { MetadataBlockInfoMockRepository } from './MetadataBlockInfoMockRepository'
 import { MetadataBlockInfoMockLoadingRepository } from './MetadataBlockInfoMockLoadingRepository'
 import { MetadataBlockInfoMockErrorRepository } from './MetadataBlockInfoMockErrorRepository'
+import { NotImplementedModalProvider } from '../../sections/not-implemented/NotImplementedModalProvider'
 
-const meta: Meta<typeof CreateDatasetForm> = {
+const meta: Meta<typeof CreateDataset> = {
   title: 'Pages/Create Dataset',
-  component: CreateDatasetForm,
-  decorators: [WithI18next, WithLayout]
+  component: CreateDataset,
+  decorators: [WithI18next, WithLayout],
+  parameters: {
+    // Sets the delay for all stories.
+    chromatic: { delay: 15000, pauseAnimationAtEnd: true }
+  }
 }
 export default meta
-type Story = StoryObj<typeof CreateDatasetForm>
+type Story = StoryObj<typeof CreateDataset>
 
 export const Default: Story = {
   render: () => (
-    <CreateDatasetForm
-      repository={new DatasetMockRepository()}
-      metadataBlockInfoRepository={new MetadataBlockInfoMockRepository()}
-    />
+    <NotImplementedModalProvider>
+      <CreateDataset
+        repository={new DatasetMockRepository()}
+        metadataBlockInfoRepository={new MetadataBlockInfoMockRepository()}
+      />
+    </NotImplementedModalProvider>
   )
 }
 
 export const LoadingMetadataBlocksConfiguration: Story = {
   render: () => (
-    <CreateDatasetForm
+    <CreateDataset
       repository={new DatasetMockRepository()}
       metadataBlockInfoRepository={new MetadataBlockInfoMockLoadingRepository()}
     />
@@ -35,7 +42,7 @@ export const LoadingMetadataBlocksConfiguration: Story = {
 
 export const ErrorLoadingMetadataBlocksConfiguration: Story = {
   render: () => (
-    <CreateDatasetForm
+    <CreateDataset
       repository={new DatasetMockRepository()}
       metadataBlockInfoRepository={new MetadataBlockInfoMockErrorRepository()}
     />